Add SpotBugs for static analysis as FindBugs is no longer maintained
As seen on https://gleclaire.github.io/findbugs-maven-plugin/ findbugs
is no longer maintained and the recommendation is to use SpotBugs.

SpotBugs reports are supported out of the box with SonarQube. Once
generated the reports can be imported via:

```
-Dsonar.java.spotbugs.reportPaths="./target/spotbugsXml.xml"
```

(assuming the default path for SpotBugs report)

This commit includes a 'bug' (Unused field) to verify detection in the
final report generated by SonarQube